Implement first prototype of manual email send.
